feat: restructure files to use react best practice
