+++ services/web-app/src/app/api/auth/[...nextauth]/route.ts
@@ -1,4 +1,4 @@
-import { authOptions } from "@/lib/auth";
+import { authOptions } from "@/app/lib/auth";
 import NextAuth from "next-auth";
 
 const handler = NextAuth(authOptions);
