+++ services/web-app/src/app/lib/auth.ts
